requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
PyYAML==6.0.2
websockets==15.0.1
httpx==0.28.1
python-dotenv==1.0.1
edge-tts
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2